requirements.txt
fastapi
uvicorn
pycryptodome
psycopg2-binary
pymongo